Multivariate Kurtosis

Mardia's estimate of multivariate kurtosis is given by $$ b_{2, k} = \frac{1}{n} \sum_{i = 1}^{n} \left[ \left( \mathbf{X}_{i} - \mathbf{\bar{X}} \right)^{T} \boldsymbol{\hat{\Sigma}}^{-1} \left( \mathbf{X}_{i} - \mathbf{\bar{X}} \right) \right]^{2} $$ where

\(\mathbf{X}\) is the \(n \times k\) sample data

\(\mathbf{\bar{X}}\) represent sample means

\(\mathbf{X}_{i} - \mathbf{\bar{X}}\) represent deviations from the mean

\(\boldsymbol{\hat{\Sigma}}\) is the estimated variance-covariance matrix of \(\mathbf{X}\) using sample data

mardiakurt(X)

Arguments

X
Matrix or data frame.

Value

Returns a vector with the following elements

b2

Estimate of multivariate kurtosis \(\left( b_{2, k} \right)\) .

z

z-statistic \(\left( \frac{b_{2, k} - k \left( k + 2 \right)}{\sqrt{\frac{8 k \left( k + 2 \right)}{n}}} \right)\) .

p

p-value associated with the z-statistic.

Details

If the null hypothesis that \(\mathbf{X}\) comes from a multivariate normal distribution is true, \(b_{2, k}\) follows a normal distribution with a mean of \(k \left( k + 2 \right)\) and a variance of \(\frac{8 k \left( k + 2 \right)}{n}\). Consequently, $$ \frac{b_{2, k} - k \left( k + 2 \right)}{\sqrt{\frac{8 k \left( k + 2 \right)}{n}}} $$ asymptotically follows a standard normal distribution \(\mathcal{N} \left( 0, 1 \right)\) .

References

Mardia, K. V. (1970). Measures of multivariate skewness and kurtosis with applications. Biometrika, 57(3), 519-530. doi:10.2307/2334770.

Mardia, K. V. (1974). Applications of Some Measures of Multivariate Skewness and Kurtosis in Testing Normality and Robustness Studies. Sankhyā: The Indian Journal of Statistics, Series B (1960-2002), 36(2), 115-128.

Examples

set.seed(42)
n <- 100
mu <- c(0, 0, 0)
Sigma <- matrix(
  data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1),
  ncol = 3
)
X <- MASS::mvrnorm(n = n, mu = mu, Sigma = Sigma)
mardiakurt(X)

#>         b2          z          p 
#> 15.9079802  0.8288687  0.4071787